samples: bluetooth: move GATT services to subsys/bluetooth
This commit moves the BLE GATT service implementations from
samples/bluetooth/gatt to subsys/bluetooth/services.
