+++ samples/bluetooth/peripheral_csc/CMakeLists.txt
@@ -5,8 +5,8 @@ project(NONE)
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE
   ${app_sources}
-  ../gatt/dis.c
-  ../gatt/bas.c
+  $ENV{ZEPHYR_BASE}/subsys/bluetooth/services/dis.c
+  $ENV{ZEPHYR_BASE}/subsys/bluetooth/services/bas.c
   )
 
 zephyr_library_include_directories($ENV{ZEPHYR_BASE}/samples/bluetooth)
