+++ samples/bluetooth/peripheral_dis/CMakeLists.txt
@@ -5,7 +5,7 @@ project(NONE)
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE
   ${app_sources}
-  ../gatt/dis.c
+  $ENV{ZEPHYR_BASE}/subsys/bluetooth/services/dis.c
   )
 
 zephyr_library_include_directories($ENV{ZEPHYR_BASE}/samples/bluetooth)
