+++ subsys/bluetooth/services/hrs.c
